Version 1
WiFi Controlled Smart Socket.

The socket was bought from clas ohlson. The socket is 2x2 Euro socket. The PCB has four relay's along with 4 LED's. The PCB fits inside the socket housing and there is a 3d printed casing for the socket. Step file can be found in the files. 3d Casing was designed by Harri Kuittinen. https://www.linkedin.com/in/harri-kuittinen-link/

Eagle schematics and Layout is attached along with the arduino code. WiFi uC is ESP8266. This version 1 is controlled with Thinger.io which is a cloud platform. You can have 2 devices for free int he platform so just sign up and make a device and copy the device information into the program and upload.